+++ R/LinkFunc.R
@@ -7,6 +7,8 @@
 #' @encoding UTF-8
 #' @author Jaime Mosquera Gutiérrez, \email{jmosquerag@unal.edu.co}
 #' @description
+#' `r lifecycle::badge("experimental")`
+#'
 #' \code{log_link} object provides a way to implement logit link function that
 #' \code{\link{maxlogL}} needs to perform estimation. See documentation for
 #' \code{\link{maxlogL}} for further information on parameter estimation and implementation
@@ -16,6 +18,7 @@
 #' @export
 #'
 #' @examples
+#' #--------------------------------------------------------------------------------
 #' # Estimation of proportion in binomial distribution with 'logit' function
 #' # 10 trials, probability of success equals to 30%)
 #' N <- rbinom(n = 100, size = 10, prob = 0.3)
@@ -35,6 +38,8 @@
 #' ginv <- logit_link()$g_inv
 #' curve(ginv(x), from = -10, to = 10)
 #'
+#' #--------------------------------------------------------------------------------
+#'
 #' @details \code{logit_link} is part of a family of generic functions with no input arguments that
 #' defines and returns a list with details of the link function:
 #' \enumerate{
@@ -52,10 +57,10 @@ logit_link <- function(){
   name <- "logit"
   g <- function(eta) log(eta/(1 - eta))
   g_inv <- function(eta) exp(eta)/(exp(eta) + 1)
-  dg.eta <- function(eta) 1/(eta*(1 - eta))
+  dg.eta <- function(eta) pmax(1/(eta*(1 - eta)), .Machine$double.eps)
   dg_inv.eta <- function(eta) exp(eta)/(exp(eta) + 1)^2
-  out <- list(name = name, g = g, g_inv = g_inv,
-              dg.eta = dg.eta, dg_inv.eta = dg_inv.eta)
+  out <- structure(list(name = name, g = g, g_inv = g_inv,
+                        dg.eta = dg.eta, dg_inv.eta = dg_inv.eta))
   return(out)
 }
 #=============================================================================#
@@ -106,11 +111,11 @@ logit_link <- function(){
 log_link <- function(){
   name <- "log"
   g <- function(eta) log(eta)
-  g_inv <- function(eta) exp(eta)
+  g_inv <- function(eta) pmax(exp(eta), .Machine$double.eps)
   dg.eta <- function(eta) 1/eta
-  dg_inv.eta <- function(eta) exp(eta)
-  out <- list(name = name, g = g, g_inv = g_inv,
-              dg.eta = dg.eta, dg_inv.eta = dg_inv.eta)
+  dg_inv.eta <- function(eta) pmax(exp(eta), .Machine$double.eps)
+  out <- structure(list(name = name, g = g, g_inv = g_inv,
+                        dg.eta = dg.eta, dg_inv.eta = dg_inv.eta))
   return(out)
 }
 #=============================================================================#
@@ -164,8 +169,8 @@ NegInv_link <- function(){
   g_inv <- function(eta) -1/eta
   dg.eta <- function(eta) 1/(eta*eta)
   dg_inv.eta <- function(eta) 1/(eta*eta)
-  out <- list(name = name, g = g, g_inv = g_inv,
-              dg.eta = dg.eta, dg_inv.eta = dg_inv.eta)
+  out <- structure(list(name = name, g = g, g_inv = g_inv,
+                        dg.eta = dg.eta, dg_inv.eta = dg_inv.eta))
   return(out)
 }
 #=============================================================================#
@@ -175,7 +180,7 @@ InvAdd_link <- function(){
   g_inv <- function(eta) -eta
   dg.eta <- function(eta) -1
   dg_inv.eta <- function(eta) -1
-  out <- list(name = name, g = g, g_inv = g_inv,
-              dg.eta = dg.eta, dg_inv.eta = dg_inv.eta)
+  out <- structure(list(name = name, g = g, g_inv = g_inv,
+                        dg.eta = dg.eta, dg_inv.eta = dg_inv.eta))
   return(out)
 }